examples/distributed/social-media/tests/requirements.txt
# Testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-timeout==2.2.0
pytest-xdist==3.5.0  # For parallel test execution

# HTTP clients
httpx==0.25.2
requests==2.31.0

# Docker and orchestration
docker==6.1.3
docker-compose==1.29.2

# OpenTelemetry for trace analysis
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-httpx==0.42b0

# Data generation and validation
faker==20.1.0
pydantic==2.5.2

# Metrics and monitoring
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3  # For retry logic
colorama==0.4.6  # For colored output
tabulate==0.9.0  # For report formatting

# Chaos engineering
toxiproxy-python==0.1.0  # For network chaos simulation
